+++ internal/service/security_control/resource_automated_action.go
@@ -17,7 +17,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/booldefault"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/tfsdk"
-	iaas "github.com/sacloud/iaas-api-go"
+	"github.com/sacloud/saclient-go"
 	seccon "github.com/sacloud/security-control-api-go"
 	v1 "github.com/sacloud/security-control-api-go/apis/v1"
 	"github.com/sacloud/terraform-provider-sakura/internal/common"
@@ -220,7 +220,7 @@ func getAutomatedAction(ctx context.Context, client *v1.Client, id string, state
 	erOp := seccon.NewAutomatedActionsOp(client)
 	aa, err := erOp.Read(ctx, id)
 	if err != nil {
-		if iaas.IsNotFoundError(err) {
+		if saclient.IsNotFoundError(err) {
 			state.RemoveResource(ctx)
 			return nil
 		}